feat: skip-nix-install
We have some problems in our custom runner with this action so
I've tried it with skipping the installation via the included action and
it works in our runners (we do not use the GitHub provided ones, we are
using custom runners).

In case you are wondering which action we are using,
we use: `cachix/install-nix-action@v23` - I guess it has something to do
with the usage of `sudo` in the other action but I'm not 100% sure on
that.