feat: skip-nix-install (#23)
We have some problems in our custom runner with this action so I've
tried it with skipping the installation via the included action and it
works in our runners (we do not use the GitHub provided ones, we are
using custom runners).

In case you are wondering which action we are using, we use:
`cachix/install-nix-action@v23` - I guess it has something to do with
the usage of `sudo` in the other action but I'm not 100% sure on that.
I'm also not sure about the differences between these two actions so
I've opted for the possibility to skip the installation rather than
changing the action as I guess you've put a lot of thought into which
action to use.

Edit: I need to correct myself, it's not because of `sudo` but because
of `systemd`. I tried to run the action without `systemd` by passing in
`init: none` and `planner: linux-multi` but I got other errors. So I
would prefer to be able to just skip the installation and use my own / a
different action to install `nix`.