Replace github cache action with restore and save (#24)
### What
Instead of letting `action/cache` do the restoring and saving cache for
us, we choose when to restore and when to save cache.

In this case, we will be saving cache immediately after the devbox
package installation, instead of waiting for the action post run. This
way, the `tar` sudo permission stays intact until the cache is saved.

More specifically, `action/cache/save` runs before the last step of
restoring `tar` to its original permission (as oppose to after)