Add source-ref input to build devbox from source
Its sometimes useful to install a version of devbox that hasn't been
tagged for a release yet. Add a `source-ref` input to make the action
clone and build devbox from source instead of using the installer
script.

This is intended for development and debugging. It should not be used in
production workflows.

Example:

```yaml
- uses: jetify-com/devbox-install-action@main
  with:
    source-ref: main
```

The cloned devbox repository and the built binary are placed in a
.devbox-installer-action directory in the GitHub workspace.

If Go is already in the PATH, the action won't install it again. This
gives the caller an opportunity to install Go themselves and specify the
Go version or caching options.

For now, caching the CLI is always disabled when building from source.