feat(action): Add extra-nix-config input for custom nix.conf settings
# Issue

See #74:

When running actions on self-hosted GitHub Enterprise (GHES) installations, the `github.token` provided is specific to the private GHES instance. This token cannot be used to access resources on `github.com`. While the existing configuration option `disable-nix-access-token` allows users to disable the usage of `github.token` as a Nix access token, this workaround fails once the unauthenticated rate limit for `api.github.com` is exceeded.

Currently, there is no mechanism to configure a custom, working Nix access token for such scenarios. This limitation creates challenges for GHES users who need to:
- Authenticate against `api.github.com` without exceeding rate limits.
- Configure Nix access tokens for other APIs, such as private hosts, when using Nix packages from private repositories.

# Fix

Add a new configuration option `extra-nix-config` that gets appended to `nix.conf` if passed.

This can be used to configure access tokens, and I added this use case as an example to the README.md.